♻️ refactor : ResponsiveLayout 컴포넌트 삭제 및 코드 이동
/mypage/layout.tsx로 코드 이동

Resolves : #106
